+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 
 ### Added
 
+* `OpenSSLDialer` dialer factory.
+
 ### Changed
 
 ### Fixed